Keep only a single copy of the current 'initial schema' in source, move old 'initial schemas' to test-only files under tests/resources/db
