+++ mlflow/store/sqlalchemy_store.py
@@ -23,7 +23,7 @@ from mlflow.utils.file_utils import mkdir, local_file_uri_to_path
 from mlflow.utils.validation import _validate_batch_log_limits, _validate_batch_log_data,\
     _validate_run_id, _validate_metric
 from mlflow.store.db.utils import _upgrade_db, _get_alembic_config, _stamp_initial_db
-from mlflow.store.dbmodels.initial_valid_models import Base as InitialBase
+from mlflow.store.dbmodels.initial_models import Base as InitialBase
 
 
 _logger = logging.getLogger(__name__)
